Provide a tool to initialize super block and super wallet: $ cd ${project_home}/test $ make $ ./init_super_wallet.sh

+++ server/FPProto.h
@@ -96,8 +96,8 @@ public:
     string _id;
     string _from;
     string _to;
-    int32_t _value;
-    int32_t _timestamp;
+    int64_t _value;
+    uint64_t _timestamp;
     //vector<Confirm> _confirmList;
 
     template <class Archive>
